requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
pypdf
python-dotenv
requests
streamlit
faiss-cpu
langchain-huggingface 